refactor(ao): extract agents reference scanner
