+++ cli/internal/adapters/agentsreferences/references_test.go
@@ -1,5 +1,5 @@
 // practices: [wiki-knowledge-surface, design-by-contract]
-package main
+package agentsreferences
 
 import (
 	"encoding/json"
@@ -8,19 +8,101 @@ import (
 	"path/filepath"
 	"sort"
 	"testing"
+
+	"github.com/boshu2/agentops/cli/internal/adapters/agentsurface"
 )
 
-// TestAgentsWriteSurfaces_GoShellScannerParity asserts that the Go scanner
-// (scanProductionAgentsReferences in agents_smoke_test.go) and the shell
-// scanner (scripts/check-agents-write-surfaces.sh) detect the same set of
-// `.agents/<X>` references when given the same fixture tree.
-//
-// Closes finding f-2026-04-27-004: "Duplicated contract scanners drift when
-// a new accepted syntax is added to one validator without paired fixture
-// coverage in the mirror validator." The fixture covers every recognized
-// syntax form (literal in .go, filepath.Join in .go, literal in .sh,
-// literal in .bash) plus one form each scanner should reject (test files,
-// non-production directories, false-prefix matches).
+// TestAgentsWriteSurfaces_EachAllowlistEntryHasProductionReference is the
+// read-side counterpart to scripts/check-agents-write-surfaces.sh. The shell
+// script flags production code that references an undocumented surface; this
+// test flags catalogued surfaces that have no production reference.
+func TestAgentsWriteSurfaces_EachAllowlistEntryHasProductionReference(t *testing.T) {
+	repoRoot, contractPath := findContractRoot(t)
+	contractData, err := os.ReadFile(contractPath)
+	if err != nil {
+		t.Fatalf("read contract %s: %v", contractPath, err)
+	}
+
+	allowlist := agentsurface.ParseAllowlist(string(contractData))
+	if len(allowlist) == 0 {
+		t.Fatal("allowlist parsed empty: contract markers missing or malformed")
+	}
+
+	refs, err := ScanProduction(repoRoot)
+	if err != nil {
+		t.Fatalf("scanning production refs: %v", err)
+	}
+
+	for _, entry := range allowlist {
+		t.Run(entry, func(t *testing.T) {
+			if !refs[entry] {
+				t.Errorf(
+					"allowlist entry %q has no production reference under cli/, scripts/, hooks/, or lib/. "+
+						"Either add a write site or remove %q from docs/contracts/agents-write-surfaces.md.",
+					entry, entry,
+				)
+			}
+		})
+	}
+}
+
+func TestScanProduction_FindsKnownLiteral(t *testing.T) {
+	tmp := t.TempDir()
+	if err := os.MkdirAll(filepath.Join(tmp, "cli", "cmd", "ao"), 0o755); err != nil {
+		t.Fatal(err)
+	}
+	if err := os.MkdirAll(filepath.Join(tmp, "scripts"), 0o755); err != nil {
+		t.Fatal(err)
+	}
+
+	goSrc := []byte(`package main
+
+import "path/filepath"
+
+const _ = ".agents/learnings/foo.md"
+var _ = filepath.Join(cwd, ".agents", "packets", "promoted")
+var _ = filepath.Join(".agents", "wiki", "sources")
+`)
+	if err := os.WriteFile(filepath.Join(tmp, "cli", "cmd", "ao", "thing.go"), goSrc, 0o644); err != nil {
+		t.Fatal(err)
+	}
+
+	goTest := []byte(`package main
+const _ = ".agents/test-only-surface/x"
+`)
+	if err := os.WriteFile(filepath.Join(tmp, "cli", "cmd", "ao", "thing_test.go"), goTest, 0o644); err != nil {
+		t.Fatal(err)
+	}
+
+	sh := []byte("#!/usr/bin/env bash\necho .agents/releases/run.json\n")
+	if err := os.WriteFile(filepath.Join(tmp, "scripts", "ship.sh"), sh, 0o644); err != nil {
+		t.Fatal(err)
+	}
+
+	refs, err := ScanProduction(tmp)
+	if err != nil {
+		t.Fatalf("scan: %v", err)
+	}
+	if !refs["learnings"] {
+		t.Error("expected 'learnings' from production go file")
+	}
+	if !refs["packets"] {
+		t.Error("expected 'packets' from filepath.Join production go file")
+	}
+	if !refs["wiki"] {
+		t.Error("expected 'wiki' from filepath.Join in production go file")
+	}
+	if !refs["releases"] {
+		t.Error("expected 'releases' from shell script")
+	}
+	if refs["test-only-surface"] {
+		t.Error("test files must be excluded; 'test-only-surface' should not appear")
+	}
+}
+
+// TestAgentsWriteSurfaces_GoShellScannerParity asserts that the Go scanner and
+// the shell scanner detect the same set of `.agents/<X>` references when given
+// the same fixture tree.
 func TestAgentsWriteSurfaces_GoShellScannerParity(t *testing.T) {
 	repoRoot := findRepoRootForParity(t)
 	scriptPath := filepath.Join(repoRoot, "scripts", "check-agents-write-surfaces.sh")
@@ -30,7 +112,7 @@ func TestAgentsWriteSurfaces_GoShellScannerParity(t *testing.T) {
 
 	tmp := buildParityFixture(t)
 
-	goRefs, err := scanProductionAgentsReferences(tmp)
+	goRefs, err := ScanProduction(tmp)
 	if err != nil {
 		t.Fatalf("Go scanner error: %v", err)
 	}
@@ -44,16 +126,34 @@ func TestAgentsWriteSurfaces_GoShellScannerParity(t *testing.T) {
 			diff(want, got), diff(got, want), want, got)
 	}
 
-	// Pin the expected fixture set so a future fixture edit that shrinks
-	// either scanner without touching the other is caught here too.
 	expected := []string{"learnings", "memory", "packets", "releases", "wiki"}
 	if !equalStringSets(want, expected) {
 		t.Fatalf("fixture coverage drifted: scanners saw %v, expected %v", want, expected)
 	}
 }
 
-// buildParityFixture writes a synthetic repo layout containing every
-// recognized syntax form plus the contract doc the shell script needs.
+func findContractRoot(t *testing.T) (string, string) {
+	t.Helper()
+	cwd, err := os.Getwd()
+	if err != nil {
+		t.Fatalf("getwd: %v", err)
+	}
+	dir := cwd
+	for i := 0; i < 8; i++ {
+		probe := filepath.Join(dir, "docs", "contracts", "agents-write-surfaces.md")
+		if _, err := os.Stat(probe); err == nil {
+			return dir, probe
+		}
+		parent := filepath.Dir(dir)
+		if parent == dir {
+			break
+		}
+		dir = parent
+	}
+	t.Fatalf("could not find docs/contracts/agents-write-surfaces.md walking up from %s", cwd)
+	return "", ""
+}
+
 func buildParityFixture(t *testing.T) string {
 	t.Helper()
 	tmp := t.TempDir()
@@ -65,8 +165,6 @@ func buildParityFixture(t *testing.T) string {
 	mustMkdir(t, filepath.Join(tmp, "docs", "contracts"))
 	mustMkdir(t, filepath.Join(tmp, "skills"))
 
-	// Production .go: literal + filepath.Join. Test file (excluded) and
-	// a similar-looking false prefix to confirm both scanners reject it.
 	goSrc := []byte(`package main
 
 import "path/filepath"
@@ -83,15 +181,11 @@ const _ = ".agents/test-only-surface/x"
 `)
 	mustWrite(t, filepath.Join(tmp, "cli", "cmd", "ao", "thing_test.go"), goTest)
 
-	// Production shell: .sh and .bash both recognized.
 	mustWrite(t, filepath.Join(tmp, "scripts", "ship.sh"),
 		[]byte("#!/usr/bin/env bash\necho .agents/releases/run.json\n"))
 	mustWrite(t, filepath.Join(tmp, "lib", "store.bash"),
 		[]byte("#!/usr/bin/env bash\nstore=.agents/memory/cache\n"))
 
-	// Contract doc: minimal typed table plus an intentionally narrow allowlist.
-	// The parity test only cares about the referenced set; allowlist and
-	// classification failures are expected and tolerated as exit 1.
 	contract := []byte(`# Test fixture
 
 | Surface | Lifecycle | Allowed writers | Mutation lane | Purpose |
@@ -111,9 +205,6 @@ func runShellScanner(t *testing.T, scriptPath, repoRoot string) map[string]bool 
 	t.Helper()
 	cmd := exec.Command("bash", scriptPath, "--json")
 	cmd.Env = append(os.Environ(), "AGENTS_WRITE_SURFACES_REPO_ROOT="+repoRoot)
-	// --json writes payload to stdout even when exit=1 (undocumented
-	// surfaces present, expected here because the fixture's allowlist is
-	// intentionally empty). Capture stdout directly and tolerate exit=1.
 	out, err := cmd.Output()
 	if err != nil {
 		exitErr, ok := err.(*exec.ExitError)
@@ -140,9 +231,6 @@ func runShellScanner(t *testing.T, scriptPath, repoRoot string) map[string]bool 
 
 func findRepoRootForParity(t *testing.T) string {
 	t.Helper()
-	// Reuse findContractRoot from agents_smoke_test.go — it walks up to
-	// docs/contracts/agents-write-surfaces.md, which is exactly the layout
-	// we need.
 	root, _ := findContractRoot(t)
 	return root
 }
